Oracle data is now pulled from API
